All Codes/Regression/boston/requirements.txt
pandas
matplotlib.pyplot
seaborn
scikitlearn
numpy